Move make_vcz to test utils and use in more tests
